rpgmz_tool::command::commands2py::command::code

Struct CommandCode

pub struct CommandCode(pub u32);

A command code

Tuple Fields
0: u32
Implementations

impl CommandCode

pub const NOP: Self

This has no implementation.

pub const SHOW_TEXT: Self

pub const SHOW_CHOICES: Self

pub const COMMENT: Self

pub const CONDITONAL_BRANCH: Self

pub const LOOP: Self

pub const COMMON_EVENT: Self

pub const LABEL: Self

pub const JUMP_TO_LABEL: Self

pub const CONTROL_SWITCHES: Self

pub const CONTROL_VARIABLES: Self

pub const CONTROL_SELF_SWITCH: Self

pub const CHANGE_ITEMS: Self

pub const CHANGE_ARMORS: Self

pub const CHANGE_PARTY_MEMBER: Self

pub const CHANGE_SAVE_ACCESS: Self

pub const TRANSFER_PLAYER: Self

pub const SET_EVENT_LOCATION: Self

pub const SET_MOVEMENT_ROUTE: Self

pub const SHOW_ANIMATION: Self

pub const SHOW_BALLOON_ICON: Self

pub const FADEOUT_SCREEN: Self

pub const FADEIN_SCREEN: Self

pub const SHAKE_SCREEN: Self

pub const WAIT: Self

pub const ERASE_PICTURE: Self

pub const PLAY_BGM: Self

pub const FADEOUT_BGM: Self

pub const PLAY_SE: Self

pub const BATTLE_PROCESSING: Self

pub const NAME_INPUT_PROCESSING: Self

pub const CHANGE_HP: Self

pub const CHANGE_MP: Self

pub const GAME_OVER: Self

pub const RETURN_TO_TITLE_SCREEN: Self

pub const PLUGIN_COMMAND: Self

pub const TEXT_DATA: Self

pub const WHEN: Self

pub const WHEN_END: Self

I think this is an end for the WHEN block. I can’t be sure as the game doesn’t actually care if this exists; it just ignores it, only taking into account indents.

pub const COMMENT_EXTRA: Self

pub const ELSE: Self

pub const CONDITONAL_BRANCH_END: Self

I think this is an end for the CONDITONAL_BRANCH block. I can’t be sure as the game doesn’t actually care if this exists; it just ignores it, only taking into account indents.

pub const REPEAT_ABOVE: Self

This is an end for the LOOP command. In contrast with the other block ends, most of logic is here. All this does is reset the command index to the top LOOP marker command.

pub const SET_MOVEMENT_ROUTE_EXTRA: Self

This is likely related to move routes somehow, like how the TEXT_DATA command extends the SHOW_TEXT command. However, I can’t find the implementation of this instruction. Furthermore, I don’t know why it’s even included; Its data always duplicates the data of the SET_MOVEMENT_ROUTE command.

pub const IF_WIN: Self

pub const IF_ESCAPE: Self

pub const IF_LOSE: Self

pub const BATTLE_RESULT_END: Self

I think this is an end for an IF_WIN, IF_ESCAPE, or IF_LOSE block. I can’t be sure as the game doesn’t actually care if this exists; it just ignores it, only taking into account indents.

pub const PLUGIN_COMMAND_EXTRA: Self

This is likely related to plugin commands somehow, like how the TEXT_DATA command extends the SHOW_TEXT command. However, I can’t find the implementation of this instruction. Furthermore, I don’t know why it’s even included; Its data always duplicates the data of the PLUGIN_COMMAND args data.

pub fn as_str(self) -> Option<&'static str>